+++ src/jinja2-2/collectors/__init__.py
@@ -1,0 +1,2 @@
+# Phytoscope Collectors Package
+# 各模块的数据收集函数，每个模块返回 dict 供 Jinja2 模板渲染